refactor(76): code refactor
