--- src/data/PCDetailsData.ts
@@ -1,14 +1,0 @@
-import { TPCDetails } from "../types/PC/TPCDetails";
-
-export const pcDetails: TPCDetails = {
-    id: "83305d70-4a77-4811-a365-eeda3aa5f5fd",
-    processorBrand: "Intel",
-    processorName: "Intel Core i7 9gen 9700 2.9 - 4.0 GHz",
-    graphicsCardBrand: "NVIDIA",
-    graphicsCardName: "NVIDIA GeForce RTX 2090",
-    ramGBCapacity: 64,
-    driveGBCapacity: 256,
-    driveType: "SSD",
-    operatingSystem: "Windows",
-    price: 1049.99,
-};
