--- src/types/PC/TCheckboxIDName.ts
@@ -1,6 +1,0 @@
-import { TIDNameType } from "./TPCComboData";
-
-export type TCheckboxIDName = {
-    name: string;
-    options?: TIDNameType[] | null;
-};
