--- src/types/PC/TCheckboxName.ts
@@ -1,4 +1,0 @@
-export type TCheckboxName = {
-    name: string;
-    options?: string[] | null;
-};
