--- src/types/PC/TPCFilter.ts
@@ -1,19 +1,0 @@
-import { UUID } from "crypto";
-
-export type TPCFilter = {
-    processorBrands: UUID[];
-    graphicsCardBrands: UUID[];
-    ramCapacities: string[];
-    driveCapacities: string[];
-    driveTypes: string[];
-    operatingSystems: UUID[];
-    priceFrom: number;
-    priceTo: number | null;
-    pagingAndSortingRequest: TPagingAndSortingRequest;
-};
-
-type TPagingAndSortingRequest = {
-    pageNumber: number;
-    pageSize: number;
-    ascendingFlag: boolean | null;
-};
