--- src/types/PC/TPCSimple.ts
@@ -1,9 +1,0 @@
-import { UUID } from "crypto";
-
-export type TPCSimple = {
-    id: UUID;
-    processorName: string;
-    graphicsCardName: string;
-    ramCapacity: number;
-    price: number;
-};
