--- src/types/PC/TPCsPageResponse.ts
@@ -1,13 +1,0 @@
-import { TPCSimple } from "./TPCSimple";
-
-export type TPCsPageResponse = {
-    pcs: TPCSimple[];
-    pagingAndSortingMetadata: TPagingAndSortingMetadata;
-};
-
-type TPagingAndSortingMetadata = {
-    pageNumber: number;
-    pageSize: number;
-    pagesCount: number;
-    elementsCount: number;
-};
