--- src/types/PC/TPCsProps.ts
@@ -1,3 +1,0 @@
-export type TPCsProps = {
-    userRole: string;
-};
