--- src/types/TPagePagination.ts
@@ -1,8 +1,0 @@
-export type TPagePagination = {
-    pagesCount: number;
-    pageNumber: number;
-    pageSize: number;
-    onChangePage: (pageNumber: number, pageSize: number) => void;
-};
-
-export type TPageSize = Omit<TPagePagination, "pageNumber">;
